+++ mesh.c
@@ -8,18 +8,18 @@
 #define FILL_MESH_INDICES_PER_HEX 12
 #define MESH_FILL_FRACTION 0.95
 
-static vec2 mesh_hex_corner(i8 corner)
+static vec2 mesh_hex_corner(int corner)
 {
-	f64 angle = M_PI / 3.0 * (0.5 + corner);
+	double angle = M_PI / 3.0 * (0.5 + corner);
 	return (vec2) {
 		cos(angle) * MESH_FILL_FRACTION,
 		sin(angle) * MESH_FILL_FRACTION,
 	};
 }
 
-void fill_mesh_initialize(FillMesh *m, i32 size_c, i32 size_r)
+void fill_mesh_initialize(FillMesh *m, int size_c, int size_r)
 {
-	i32 num_hexes = size_c * size_r;
+	int num_hexes = size_c * size_r;
 	m->vertices_length = FILL_MESH_VERTICES_PER_HEX * num_hexes;
 	m->indices_length = FILL_MESH_INDICES_PER_HEX * num_hexes;
 	m->size_c = size_c;
@@ -45,8 +45,8 @@ void fill_mesh_initialize(FillMesh *m, i32 size_c, i32 size_r)
 	};
 
 	Hex h;
-	i32 vi = 0;
-	i32 ii = 0;
+	int vi = 0;
+	int ii = 0;
 
 	for (h.r = 0; h.r < size_r; ++h.r) {
 		for (h.c = 0; h.c < size_c; ++h.c) {
@@ -54,13 +54,13 @@ void fill_mesh_initialize(FillMesh *m, i32 size_c, i32 size_r)
 					hex_to_vec(
 					hex_from_storage(h)));
 
-			for (i32 i = 0; i < FILL_MESH_INDICES_PER_HEX; ++i) {
+			for (int i = 0; i < FILL_MESH_INDICES_PER_HEX; ++i) {
 				m->indices[ii + i] = vi + indices_single[i];
 			}
 
 			ii += FILL_MESH_INDICES_PER_HEX;
 
-			for (i32 i = 0; i < FILL_MESH_VERTICES_PER_HEX; ++i) {
+			for (int i = 0; i < FILL_MESH_VERTICES_PER_HEX; ++i) {
 				FillMeshVertex *vx = &m->vertices[vi + i];
 				vx->x = corners[i].x + center.x;
 				vx->y = corners[i].y + center.y;
@@ -79,14 +79,14 @@ void fill_mesh_terminate(FillMesh *m)
 	free(m->indices);
 }
 
-void instance_mesh_initialize(InstanceMesh *m, i32 length)
+void instance_mesh_initialize(InstanceMesh *m, int length)
 {
 	m->vertices = malloc(length * sizeof *m->vertices);
 	m->vertices_length = length;
 	m->vertices_capacity = length;
 }
 
-void instance_mesh_resize(InstanceMesh *m, i32 length)
+void instance_mesh_resize(InstanceMesh *m, int length)
 {
 	if (length > m->vertices_capacity) {
 		instance_mesh_terminate(m);
